Split TextNode out of main file renderer
