+++ packages/solid/src/reconciler.ts
@@ -13,122 +13,11 @@ import {
 } from "@opentui/core"
 import { createRenderer } from "solid-js/universal"
 import { elements, type Element } from "./elements"
+import { insertTextNode, removeTextNode, TextNode } from "./elements/text-node"
 import { getNextId } from "./utils/id-counter"
+import { log } from "./utils/log"
 
-const GHOST_NODE_TAG = "text-ghost" as const
-
-class TextNode {
-  id: string
-  chunk: TextChunk
-  parent?: Renderable
-  textParent?: TextRenderable
-
-  constructor(chunk: TextChunk) {
-    this.id = getNextId("text-node")
-    this.chunk = chunk
-  }
-}
-
-const ChunkToTextNodeMap = new WeakMap<TextChunk, TextNode>()
-
-type DomNode = Renderable | TextNode
-
-const log = (...args: any[]) => {
-  console.log("[Reconciler]", ...args)
-}
-
-function getOrCreateTextGhostNode(parent: Renderable, anchor?: DomNode | null): TextRenderable {
-  if (anchor instanceof TextNode && anchor.textParent) {
-    // prepend text to anchor
-    return anchor.textParent
-  }
-  const children = parent.getChildren()
-
-  if (anchor instanceof Renderable) {
-    const anchorIndex = children.findIndex((el) => el.id === anchor.id)
-    const beforeAnchor = children[anchorIndex - 1]
-    if (beforeAnchor instanceof TextRenderable && beforeAnchor.id.startsWith(GHOST_NODE_TAG)) {
-      // append text to previous
-      return beforeAnchor
-    }
-  }
-
-  const lastChild = children.at(-1)
-  if (lastChild instanceof TextRenderable && lastChild.id.startsWith(GHOST_NODE_TAG)) {
-    // Append text to last child if exists
-    return lastChild
-  }
-
-  // Create a new ghost node
-  const ghostNode = new TextRenderable(getNextId(GHOST_NODE_TAG), {})
-
-  _insertNode(parent, ghostNode, anchor)
-
-  return ghostNode
-}
-
-function insertTextNode(parent: DomNode, node: TextNode, anchor?: DomNode | null): void {
-  if (!(parent instanceof Renderable)) {
-    console.warn("Attaching text node to parent text node, impossible")
-    return
-  }
-  log("Inserting text node:", node.id, "into parent:", parent.id, "with anchor:", anchor?.id)
-
-  let textParent: TextRenderable
-  // get parent text renderable
-  if (!(parent instanceof TextRenderable)) {
-    textParent = getOrCreateTextGhostNode(parent, anchor)
-  } else {
-    textParent = parent
-  }
-
-  node.textParent = textParent
-  let styledText = textParent.content
-
-  if (anchor && anchor instanceof TextNode) {
-    const anchorIndex = styledText.chunks.indexOf(anchor.chunk)
-    if (anchorIndex == -1) {
-      console.log("anchor not found")
-      return
-    }
-    styledText = styledText.insert(node.chunk, anchorIndex)
-  } else {
-    const firstChunk = textParent.content.chunks[0]
-    // Handles the default unlinked chunk
-    if (firstChunk && !ChunkToTextNodeMap.has(firstChunk)) {
-      styledText = styledText.replace(node.chunk, firstChunk)
-    } else {
-      styledText = styledText.insert(node.chunk)
-    }
-  }
-  textParent.content = styledText
-  node.parent = parent
-  return
-}
-
-function removeTextNode(parent: DomNode, node: TextNode): void {
-  if (!(parent instanceof Renderable)) {
-    // cleanup orphaned node
-    ChunkToTextNodeMap.delete(node.chunk)
-    return
-  }
-  if (parent === node.textParent && parent instanceof TextRenderable) {
-    ChunkToTextNodeMap.delete(node.chunk)
-    parent.content = parent.content.remove(node.chunk)
-  } else if (node.textParent) {
-    // check to remove ghost node
-    ChunkToTextNodeMap.delete(node.chunk)
-    let styledText = node.textParent.content
-    styledText = styledText.remove(node.chunk)
-
-    if (styledText.chunks.length > 0) {
-      node.textParent.content = styledText
-    } else {
-      node.parent?.remove(node.textParent.id)
-      node.textParent.destroyRecursively()
-    }
-  }
-}
+export type DomNode = Renderable | TextNode
 
 function _insertNode(parent: DomNode, node: DomNode, anchor?: DomNode | null): void {
   log("Inserting node:", node.id, "into parent:", parent.id, "with anchor:", anchor?.id)
@@ -199,7 +88,6 @@ export const {
             plainText: `${value}`,
           }
     const textNode = new TextNode(chunk)
-    ChunkToTextNodeMap.set(chunk, textNode)
     return textNode
   },
 
@@ -211,18 +99,7 @@ export const {
       text: new TextEncoder().encode(value),
       plainText: value,
     }
-
-    const textParent = textNode.textParent
-    if (!textParent) {
-      log("No parent found for text node:", textNode.id)
-      return
-    }
-    if (textParent instanceof TextRenderable) {
-      textParent.content = textParent.content.replace(newChunk, textNode.chunk)
-
-      textNode.chunk = newChunk
-      ChunkToTextNodeMap.set(newChunk, textNode)
-    }
+    textNode.replaceText(newChunk)
   },
 
   setProperty(node: DomNode, name: string, value: any, prev: any): void {
@@ -358,7 +235,7 @@ export const {
     if (node instanceof TextRenderable) {
       const chunk = node.content.chunks[0]
       if (chunk) {
-        return ChunkToTextNodeMap.get(chunk)
+        return TextNode.getTextNodeFromChunk(chunk)
       } else {
         return undefined
       }
@@ -402,7 +279,7 @@ export const {
           return undefined
         }
 
-        return ChunkToTextNodeMap.get(nextSibling)
+        return TextNode.getTextNodeFromChunk(nextSibling)
       }
       console.warn("Text parent is not a text node:", node.id)
       return undefined
@@ -435,7 +312,7 @@ const insertStyledText = (parent: any, value: any, current: any, marker: any) =>
   if (current) {
     if (typeof current === "object" && "__isChunk" in current) {
       // log("[Reconciler] Removing current:", current);
-      const node = ChunkToTextNodeMap.get(current)
+      const node = TextNode.getTextNodeFromChunk(current)
       if (node) {
         // log("[Reconciler] Removing chunk:", current.text);
         _removeNode(parent, node)
@@ -443,7 +320,7 @@ const insertStyledText = (parent: any, value: any, current: any, marker: any) =>
     } else if (current instanceof StyledText) {
       // log("[Reconciler] Removing current:", current);
       for (const chunk of current.chunks) {
-        const chunkNode = ChunkToTextNodeMap.get(chunk)
+        const chunkNode = TextNode.getTextNodeFromChunk(chunk)
         if (!chunkNode) continue
         // log("[Reconciler] Removing styled text:", chunk.text);
         _removeNode(parent, chunkNode)
