+++ packages/solid/src/utils/log.ts
@@ -1,0 +1,3 @@
+export const log = (...args: any[]) => {
+  console.log("[Reconciler]", ...args)
+}
